adding network connectivity check
