+++ pubspec.yaml
@@ -25,8 +25,9 @@ dependencies:
   http: ^0.12.0+2
   flutter_secure_storage: ^3.3.1+1
   intl: ^0.16.0
-  meta: ^1.1.7
   equatable: ^0.6.0
+dependency_overrides:
+  meta: ^1.1.7
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -53,8 +54,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #  - images/a_dot_burr.jpeg
+  assets:
+   - assets/images/icon.png
   #  - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
